+++ packages/emotion/src/create.ts
@@ -1,13 +1,14 @@
 import { StyleGenerator } from '@xstyled/system'
 import { createCssFunction, XCSSFunction } from './createCssFunction'
 import { createX, X } from './createX'
-import { createStyled, XStyled } from './createStyled'
+import { createStyled } from './createStyled'
 import {
   createCreateGlobalStyle,
   XCreateGlobalStyle,
 } from './createCreateGlobalStyle'
 import { createCx, Cx } from './createCx'
 import { createJsx, XJsx } from './createJsx'
+import { XStyled } from './types'
 
 interface XStyledSet<TGen extends StyleGenerator> {
   css: XCSSFunction
