+++ packages/emotion/src/index.ts
@@ -14,6 +14,7 @@ export * from './theme'
 export * from './preflight'
 export * from '@xstyled/system'
 export * from './create'
+export * from './types';
 
 // Create and export default system
 import { system } from '@xstyled/system'
